# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Blockchain Gateway Backend API
  version: 3.0.0
  description: |
    Backend API for chaincode lifecycle orchestration with RBAC.
    
    This API provides comprehensive management of:
    - User authentication and authorization with JWT tokens
    - Role-based access control (ADMIN, ORG_ADMIN, USER, VIEWER)
    - Chaincode lifecycle management (upload, validate, approve, deploy)
    - Deployment orchestration with workflow management
    - Certificate synchronization with Fabric CA
    - Audit logging for all operations
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Roles and Permissions
    
    - **ADMIN**: Full system access, can manage users, approve chaincodes, configure system
    - **ORG_ADMIN**: Organization administrator, can upload/deploy chaincodes, view users
    - **USER**: Regular user, can invoke/query chaincodes, manage assets
    - **VIEWER**: Read-only access, can query chaincodes and view assets
    
  contact:
    name: Blockchain Gateway Team
    email: support@blockchain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.blockchain.com/api/v1
    description: Production server
...
paths:
...
  /deployments/deploy:
    post:
      tags:
        - Deployment Management
      summary: Deploy chaincode
      description: Deploy a chaincode to the blockchain network
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaincodeDeployRequest'
      responses:
        '202':
          description: Deployment initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '400':
          description: Invalid deployment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments/invoke:
    post:
      tags:
        - Deployment Management
      summary: Invoke chaincode function
      description: Invoke a chaincode function (write transaction)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaincodeInvokeRequest'
      responses:
        '200':
          description: Chaincode invoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvokeResponse'
        '400':
          description: Invalid invoke data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments/query:
    post:
      tags:
        - Deployment Management
      summary: Query chaincode function
      description: Query a chaincode function (read-only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaincodeQueryRequest'
      responses:
        '200':
          description: Chaincode queried successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Invalid query data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments:
    get:
      tags:
        - Deployment Management
      summary: Get deployments list
      description: Get list of deployments with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          description: Number of records to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of records to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: status
          in: query
          description: Filter by deployment status
          required: false
          schema:
            type: string
            enum: [pending, deploying, success, failed, rolled_back]
        - name: deployed_by
          in: query
          description: Filter by deployer
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentListResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments/{deployment_id}:
    get:
      tags:
        - Deployment Management
      summary: Get deployment by ID
      description: Get deployment information by ID
      security:
        - bearerAuth: []
      parameters:
        - name: deployment_id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentDetailResponse'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deployments/{deployment_id}/status:
    get:
      tags:
        - Deployment Management
      summary: Get deployment status
      description: Get current status of a deployment
      security:
        - bearerAuth: []
      parameters:
        - name: deployment_id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployment status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentStatusResponse'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Certificate Management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        detail:
          type: string
          example: "Detailed error information"

    # Authentication schemas
...
    ChaincodeDeployRequest:
      type: object
      required:
        - chaincode_id
        - channel_name
        - target_peers
      properties:
        chaincode_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        channel_name:
          type: string
          example: "mychannel"
        target_peers:
          type: array
          items:
            type: string
          example: ["peer0.org1.example.com:7051"]
    ChaincodeInvokeRequest:
      type: object
      required:
        - chaincode_id
        - function_name
        - args
      properties:
        chaincode_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        function_name:
          type: string
          example: "CreateAsset"
        args:
          type: array
          items:
            type: string
          example: ["asset1", "100", "blue", "35", "bob"]
        channel_name:
          type: string
          default: "mychannel"
          example: "mychannel"
    ChaincodeQueryRequest:
      type: object
      required:
        - chaincode_id
        - function_name
        - args
      properties:
        chaincode_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        function_name:
          type: string
          example: "ReadAsset"
        args:
          type: array
          items:
            type: string
          example: ["asset1"]
        channel_name:
          type: string
          default: "mychannel"
          example: "mychannel"
    DeploymentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Deployment initiated"
        deployment_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          example: "pending"
    InvokeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        transaction_id:
          type: string
          example: "tx123456789"
        result:
          type: object
          additionalProperties: true
    QueryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        result:
          type: object
          additionalProperties: true
    DeploymentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            chaincode_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            channel_name:
              type: string
              example: "mychannel"
            target_peers:
              type: array
              items:
                type: string
              example: ["peer0.org1.example.com:7051"]
            deployment_status:
              type: string
              enum: [pending, deploying, success, failed, rolled_back]
              example: "success"
            deployed_by:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            deployment_date:
              type: string
              format: date-time
              nullable: true
              example: "2023-12-01T10:30:00Z"
            completion_date:
              type: string
              format: date-time
              nullable: true
              example: "2023-12-01T10:35:00Z"
            error_message:
              type: string
              nullable: true
              example: null
            deployment_logs:
              type: string
              nullable: true
              example: "Deployment completed successfully"
            created_at:
              type: string
              format: date-time
              example: "2023-12-01T10:30:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-12-01T10:35:00Z"
    DeploymentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/DeploymentDetailResponse'
        total:
          type: integer
          example: 5
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 100
    DeploymentStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            deployment_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            status:
              type: string
              enum: [pending, deploying, success, failed, rolled_back]
              example: "success"
            chaincode_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            channel_name:
              type: string
              example: "mychannel"
            target_peers:
              type: array
              items:
                type: string
              example: ["peer0.org1.example.com:7051"]
            error_message:
              type: string
              nullable: true
              example: null
            deployment_logs:
              type: string
              nullable: true
              example: "Deployment completed successfully"
            created_at:
              type: string
              format: date-time
              example: "2023-12-01T10:30:00Z"
            completion_date:
              type: string
              format: date-time
              nullable: true
              example: "2023-12-01T10:35:00Z"

    # Certificate Management schemas
...